+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added visualization Frequency graph
 - Added the Dash group visualization
+- Added the group settings for the multiline
 
 ### Changed
 - The component menu has become more compact